Started trying stuff with polices
I tried some stuff with some captures of numbers from the Roboto police. May be promising but tideous.

Updated the .glade with a top bar that displays "Sudoku solver" in bold.
Tried to resize the displayed image along with the window ; didn't work. Image doesn't shrink. So the window cannot be resized by dragging corners.