Added image treatment images to application
Now, all 6 steps of image treatment can be looked at after launching the OCR in the application.
The apply_image .c and .h files were slightly modified to redirect the saved images to a tmp folder.
Moreover, after the OCR was launched and the images loaded in memory, their files are deleted from that tmp folder.
And if the tmp folder is not present, it is created. However, that function causes a warning I can't resolve.